Added a private method to get container client, and moved interface to own file
